Use language level 6 (instead of 7).
We're doing this because in order to debug with Unity (Mono .NET
3.5) we must produce .MDB (mono debug symbols) instead of .PDB.
However, Mono 5+ only produces .PDB (specifically *portable* PDB)
which can't be converted to .MDB. So we're stuck on Mono 4, and
therefore also stuck on C# 6.

This could be resolved if the open source pdb2mdb tool were
modified to support portable PDB files.